+++ src/utils.rs
@@ -2,6 +2,6 @@
 /// available tokens in the supply.
 ///
 /// TODO: Currently placeholder, needs to be filled in once requirements known
-pub fn is_valid_amount(value: &u64) -> bool {
+pub fn is_valid_amount(value: &f64) -> bool {
     true
 }
